requirements.txt
boto3
redis
pyyaml